refactor(int): reorgaize component loading
Refs: #57
